created api.js | moved files by folders
