+++ newtab.html
@@ -4,9 +4,9 @@
 <head>
     <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
     <title>Generate standUp</title>
-    <script src="jquery.js"></script>
-    <script src="injected.js"></script>
-    <link rel="stylesheet" href="style.css">
+    <script src="scripts/jquery.js"></script>
+    <script type="module" src="scripts/popup.js"></script>
+    <link rel="stylesheet" href="styles/style.css">
 </head>
 
 <body>
